Make seperate file for label_to_key.c
