+++ srcs/_files.mk
@@ -6,7 +6,7 @@
 #    By: mmarcell <mmarcell@student.codam.nl>         +#+                      #
 #                                                    +#+                       #
 #    Created: 2020/02/28 18:56:22 by mmarcell      #+#    #+#                  #
-#    Updated: 2020/05/15 03:01:30 by nstabel       ########   odam.nl          #
+#    Updated: 2020/05/15 12:12:10 by nstabel       ########   odam.nl          #
 #                                                                              #
 # **************************************************************************** #
 
@@ -28,4 +28,5 @@ ASM_FILES :=	op.c						\
 				asm_lexical_analysis.c    	\
 				asm_translate_to_byte.c 	\
 				asm_write_translation.c		\
-				asm_is_argument.c
+				asm_is_argument.c			\
+				asm_label_to_key.c
