+++ srcs/asm_token_tab.c
@@ -6,7 +6,7 @@
 /*   By: nstabel <nstabel@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/05/12 16:40:21 by nstabel       #+#    #+#                 */
-/*   Updated: 2020/05/15 02:36:36 by nstabel       ########   odam.nl         */
+/*   Updated: 2020/05/15 12:11:53 by nstabel       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
